fix minor parsing issue; update mouse config file (#130)

+++ stac_mjx/config.py
@@ -35,7 +35,7 @@ class ModelConfig:
     SITES_TO_REGULARIZE: List[str]  # Sites to regularize during offset optimization
     RENDER_FPS: int  # FPS for rendering
     N_SAMPLE_FRAMES: int  # Number of frames to sample when computing offset residual
-    M_REG_COEF: int  # Coefficient for regularization term in offset optimization
+    M_REG_COEF: float  # Coefficient for regularization term in offset optimization
     MARKER_SIZE: float = 0.005  # Radius of site marker spheres for visualization
